Develop

+++ tests/testthat/helper-objects.R
@@ -1,39 +1,39 @@
 
-# Download the PostreSQL driver ---------------------------
-# If DATABASECONNECTOR_JAR_FOLDER exists, assume driver has been downloaded
-jarFolder <- Sys.getenv("DATABASECONNECTOR_JAR_FOLDER", unset = "")
-if (jarFolder == "") {
-  tempJarFolder <- tempfile("jdbcDrivers")
-  dir.create(tempJarFolder)
-  Sys.setenv("DATABASECONNECTOR_JAR_FOLDER" = tempJarFolder)
-  downloadJdbcDrivers("postgresql")
+# this files contains the objects used in the tests:
+if(Sys.getenv('GITHUB_ACTIONS') == 'true'){
+  # Download the PostreSQL driver ---------------------------
+  # If DATABASECONNECTOR_JAR_FOLDER exists, assume driver has been downloaded
+  jarFolder <- Sys.getenv("DATABASECONNECTOR_JAR_FOLDER", unset = "")
+  if (jarFolder == "") {
+    tempJarFolder <- tempfile("jdbcDrivers")
+    dir.create(tempJarFolder)
+    Sys.setenv("DATABASECONNECTOR_JAR_FOLDER" = tempJarFolder)
+    DatabaseConnector::downloadJdbcDrivers("postgresql")
+    
+    withr::defer({
+      unlink(tempJarFolder, recursive = TRUE, force = TRUE)
+      Sys.unsetenv("DATABASECONNECTOR_JAR_FOLDER")
+    }, testthat::teardown_env())
+  }
   
-  withr::defer({
-    unlink(tempJarFolder, recursive = TRUE, force = TRUE)
-    Sys.unsetenv("DATABASECONNECTOR_JAR_FOLDER")
-  }, testthat::teardown_env())
+  if(ifelse(is.null(Sys.info()), T, Sys.info()['sysname'] != 'Windows')){
+    # configure and activate python
+    PatientLevelPrediction::configurePython(envname = 'r-reticulate', envtype = "conda")
+    PatientLevelPrediction::setPythonEnvironment(envname = 'r-reticulate', envtype = "conda")
+    
+    # if mac install nomkl -- trying to fix github actions
+    if(ifelse(is.null(Sys.info()), F, Sys.info()['sysname'] == 'Darwin')){
+      reticulate::conda_install(envname = 'r-reticulate', packages = c('nomkl'), 
+                                forge = TRUE, pip = FALSE, pip_ignore_installed = TRUE, 
+                                conda = "auto")
+    }
+  }
 }
 
-
-# this files contains the objects used in the tests:
-travis <- T
 
 saveLoc <- tempfile("saveLoc")
 dir.create(saveLoc)
 
-
-if(ifelse(is.null(Sys.info()), T, Sys.info()['sysname'] != 'Windows')){
-  # configure and activate python
-  PatientLevelPrediction::configurePython(envname = 'r-reticulate', envtype = "conda")
-  PatientLevelPrediction::setPythonEnvironment(envname = 'r-reticulate', envtype = "conda")
-  
-  # if mac install nomkl -- trying to fix github actions
-  if(ifelse(is.null(Sys.info()), F, Sys.info()['sysname'] == 'Darwin')){
-    reticulate::conda_install(envname = 'r-reticulate', packages = c('nomkl'), 
-                              forge = TRUE, pip = FALSE, pip_ignore_installed = TRUE, 
-                              conda = "auto")
-  }
-}
 
 #@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@
 # simulated data Tests
@@ -42,9 +42,8 @@ if(ifelse(is.null(Sys.info()), T, Sys.info()['sysname'] != 'Windows')){
 data(plpDataSimulationProfile, envir = environment())
 
 # PLPDATA
-sampleSize <- 2500+sample(1000,1)
+sampleSize <- 1500+sample(300,1)
 plpData <- simulatePlpData(plpDataSimulationProfile, n = sampleSize)
-#plpData$metaData$cohortId <- plpData$metaData$cohortIds
 
 # POPULATION
 populationSettings <- createStudyPopulationSettings(
@@ -71,13 +70,33 @@ plpResult <- runPlp(
   analysisId = 'Test', 
   analysisName = 'Testing analysis',
   populationSettings = populationSettings, 
-  splitSettings = createDefaultSplitSetting(),
+  splitSettings = createDefaultSplitSetting(splitSeed = 12),
   preprocessSettings = createPreprocessSettings(), 
   modelSettings = lrSet, 
   logSettings = createLogSettings(verbosity = 'TRACE'),
   executeSettings = createDefaultExecuteSettings(), 
   saveDirectory = saveLoc
   )
+
+
+# now diagnose
+diagnoseResult <- diagnosePlp(
+  plpData = plpData,
+  outcomeId = 2,
+  analysisId = 'Test', 
+  populationSettings = populationSettings,
+  splitSettings = createDefaultSplitSetting(splitSeed = 12),
+  saveDirectory = saveLoc,
+  modelSettings = lrSet,
+  logSettings =  createLogSettings(
+    verbosity = 'DEBUG',
+    timeStamp = T,
+    logName = 'diagnosePlp Log'
+  ),
+  preprocessSettings = createPreprocessSettings(), 
+  sampleSettings = NULL, 
+  featureEngineeringSettings = NULL
+)
 
 #
 
@@ -89,39 +108,15 @@ population <- createStudyPopulation(
   )
 
 createTrainData <- function(plpData, population){
-  trainData <- list()
-  trainData$covariateData <- Andromeda::copyAndromeda(plpData$covariateData)
-  attr(trainData$covariateData, "metaData") <- attr(plpData$covariateData, "metaData")
-  trainData$labels <- population
-  trainData$folds <- data.frame(
-    rowId = population$rowId,
-    index = sample(3, nrow(population), replace = T)
-  )
-  
-  attr(trainData, "metaData")$outcomeId <- 2
-  attr(trainData, "metaData")$cohortId <- 1
-
-  class(trainData$covariateData) <- 'CovariateData'
-  
+  data <- PatientLevelPrediction::splitData(plpData = plpData, population=population,
+                                            splitSettings = PatientLevelPrediction::createDefaultSplitSetting(splitSeed = 12))
+  trainData <- data$Train
   return(trainData)
 }
 
-
-sampleSize2 <- 1000+sample(1000,1)
-plpData2 <- simulatePlpData(plpDataSimulationProfile, n = sampleSize2)
-
-population2 <- createStudyPopulation(
-  plpData = plpData2,
-  outcomeId = 2, 
-  populationSettings = populationSettings
-)
-
-sampleSizeBig <- 10000
-plpDataBig <- simulatePlpData(plpDataSimulationProfile, n = sampleSizeBig)
-
-populationBig <- createStudyPopulation(
-  plpData = plpDataBig,
-  outcomeId = 2, 
-  populationSettings = populationSettings
-)
-
+createTestData <- function(plpData, population){
+  data <- PatientLevelPrediction::splitData(plpData = plpData, population=population,
+                                            splitSettings = PatientLevelPrediction::createDefaultSplitSetting(splitSeed = 12))
+  testData <- data$Test
+  return(testData)
+}
